moved folder for rust generator up one
