--- src/generator/generators/mod.rs
@@ -1,4 +1,0 @@
-pub mod rust;
-mod types;
-
-pub use types::*;
