+++ src/generator/rust/helpers.rs
