+++ src/generator/rust/mod.rs
@@ -5,6 +5,5 @@ mod string_formatter;
 mod types;
 
 pub use filesystem::*;
-pub use generator::*;
 pub use helpers::*;
 pub use string_formatter::*;
