+++ src/generator/rust/string_formatter.rs
